refactor: :recycle: migrate storybook to its own package
