+++ package.json
@@ -5,14 +5,14 @@
   "license": "Apache-2.0",
   "scripts": {
     "example": "wrap () { pnpm -F example-\"$*\" dev | cat; }; wrap",
-    "storybook": "start-storybook -p 6006",
+    "storybook": "pnpm -F ripple-storybook start -p 6006",
     "dev": "pnpm build && concurrently \"pnpm storybook\" \"pnpm -F nuxt-app dev\"",
     "clean": "git clean -dfx && git reset --hard && pnpm",
     "build": "lerna run build --ignore docs",
-    "build:ripple": "lerna run build --scope '@dpc-sdp/ripple-ui-*' --scope @dpc-sdp/ripple-tide-api",
+    "build:ripple": "lerna run build --scope '{@dpc-sdp/ripple-ui-*,@dpc-sdp/ripple-tide-api}'",
     "build:examples": "lerna run build --scope '{nuxt-app,vue-app,webcomponents}'",
     "build:docs": "lerna run build --scope docs",
-    "build:storybook": "build-storybook",
+    "build:storybook": "pnpm -F ripple-storybook build",
     "postinstall": "patch-package&&husky install&&pnpm build:ripple",
     "preinstall": "npx only-allow pnpm",
     "release:alpha": "lerna version prerelease --preid alpha --yes",
@@ -23,8 +23,8 @@
     "test:unit": "jest --colors  --runInBand",
     "test:components-core": "pnpm -F @dpc-sdp/ripple-ui-core test:components",
     "test:components-forms": "pnpm -F @dpc-sdp/ripple-ui-forms test:components",
-    "test:storybook": "test-storybook --stories-json",
-    "test:storybook-ci": "concurrently -k -s first -n \"SB,TEST\" -c \"magenta,blue\" \"npx http-server storybook-static --port 6006 --silent\" \"wait-on tcp:6006 && pnpm test:storybook\"",
+    "test:storybook": "pnpm -F ripple-storybook test",
+    "test:storybook-ci": "pnpm -F ripple-storybook ci",
     "test:nuxt-app": "pnpm -F nuxt-app test:integration",
     "ci:nuxt-app": "pnpm -F nuxt-app test:ci",
     "start:nuxt": "pnpm -F nuxt-app start",
@@ -49,22 +49,6 @@
     "@dpc-sdp/ripple-ui-core": "workspace:2.1.0-alpha.57",
     "@dpc-sdp/ripple-ui-forms": "workspace:2.1.0-alpha.61",
     "@jest/globals": "^29.3.1",
-    "@outlinestudio/designtokenscss": "^1.0.7",
-    "@storybook/addon-a11y": "^6.5.15",
-    "@storybook/addon-actions": "^6.5.15",
-    "@storybook/addon-cssresources": "^6.2.9",
-    "@storybook/addon-essentials": "^6.5.15",
-    "@storybook/addon-interactions": "^6.5.15",
-    "@storybook/addon-jest": "^6.5.15",
-    "@storybook/addon-links": "^6.5.15",
-    "@storybook/addon-postcss": "3.0.0-alpha.1",
-    "@storybook/builder-vite": "^0.2.7",
-    "@storybook/builder-webpack4": "^6.5.15",
-    "@storybook/jest": "^0.0.10",
-    "@storybook/manager-webpack4": "^6.5.15",
-    "@storybook/test-runner": "^0.9.2",
-    "@storybook/testing-library": "^0.0.13",
-    "@storybook/vue3": "^6.5.15",
     "@testing-library/vue": "^6.6.1",
     "@types/jest": "^29.2.5",
     "@typescript-eslint/eslint-plugin": "^5.48.1",
@@ -94,7 +78,6 @@
     "postcss-normalize": "^10.0.1",
     "postcss-preset-env": "^7.8.3",
     "start-server-and-test": "^1.15.2",
-    "storybook-addon-designs": "^6.3.1",
     "ts-jest": "^29.0.4",
     "typescript": "^4.9.4",
     "unist-builder": "^3.0.0",
@@ -109,7 +92,6 @@
   },
   "dependencies": {
     "@vue/vue3-jest": "^29.2.2",
-    "concurrently": "^7.6.0",
     "rimraf": "^3.0.2",
     "vue-jest": "^3.0.7"
   },
@@ -118,6 +100,7 @@
       "@types/react": "file:./stub/types__react",
       "debug@<3.1.0": "~3.1.0",
       "glob-parent@<5.1.2": "~6.0.2",
+      "jest-mock@<28.1.0": "^28.1.0",
       "json5@<2.2.2": "~2.2.2",
       "nth-check@<2.1.1": "~2.1.1",
       "parse-url@<8.1.0": "~8.1.0",
