+++ packages/ripple-storybook/.storybook/main.js
@@ -37,23 +37,26 @@ const vitePlugins = [
   })
 ]
 module.exports = {
-  stories: ['./../packages/**/*.stories.@(js|jsx|ts|tsx|mdx)'],
+  stories: ['./../../ripple-ui*/**/*.stories.@(js|jsx|ts|tsx|mdx)'],
   framework: '@storybook/vue3',
   staticDirs: ['./public'],
   features: {
-    interactionsDebugger: true,
-    buildStoriesJson: true
+    buildStoriesJson: true,
+    storyStoreV7: false
   },
   core: {
     builder: '@storybook/builder-vite'
   },
   addons: [
-    '@storybook/addon-essentials',
+    {
+      name: '@storybook/addon-essentials',
+      options: {
+        docs: false
+      }
+    },
     '@storybook/addon-interactions',
     '@storybook/addon-jest',
-    '@storybook/addon-a11y',
-    '@outlinestudio/designtokenscss',
-    'storybook-addon-designs'
+    '@storybook/addon-a11y'
   ],
 
   async viteFinal(config, { configType }) {
