+++ packages/ripple-storybook/.storybook/preview.js
@@ -1,16 +1,13 @@
-import { RplIconSprite } from '@dpc-sdp/ripple-ui-core'
-import { registerRplFormPlugin } from '@dpc-sdp/ripple-ui-forms'
-import '@dpc-sdp/ripple-ui-core/style'
-import { withCssResources } from '@storybook/addon-cssresources'
-import { withDesign } from 'storybook-addon-designs'
 import { app } from '@storybook/vue3'
-import themes from './themes.json'
+import { registerRplFormPlugin } from '@dpc-sdp/ripple-ui-forms'
+import { RplIconSprite } from '@dpc-sdp/ripple-ui-core'
+import '@dpc-sdp/ripple-ui-core/style'
 import withBackground from './utils/withBackground'
+import svgPlaceholder from './utils/svgPlaceholder'
 // Storybook specific CSS
 import './storybook.css'
 
 // Add SVG based image placeholder for use in all stories
-import svgPlaceholder from '~/storybook/utils/svgPlaceholder'
 window.svgPlaceholder = svgPlaceholder
 
 // Ripple vue plugins
@@ -31,20 +28,13 @@ export const parameters = {
   },
   backgrounds: {
     disable: true
-  },
-  designTokensCss: {
-    label: 'Themes',
-    persistData: true,
-    themes
   }
 }
 
 export const decorators = [
-  withCssResources,
   (story) => ({
     components: { story, RplIconSprite },
     template: '<div><RplIconSprite /><story /></div>'
   }),
-  withBackground,
-  withDesign
+  withBackground
 ]
