+++ packages/ripple-storybook/.storybook/public/android-chrome-192x192.png
