+++ packages/ripple-storybook/.storybook/public/android-chrome-512x512.png
