+++ packages/ripple-storybook/.storybook/public/apple-touch-icon.png
