+++ packages/ripple-storybook/.storybook/public/favicon-32x32.png
