+++ packages/ripple-storybook/.storybook/public/favicon.ico
