+++ packages/ripple-storybook/.storybook/public/img/hero-corner-bottom.png
