+++ packages/ripple-storybook/.storybook/public/img/hero-corner-top.png
