+++ packages/ripple-storybook/.storybook/public/img/image-dark.jpg
