+++ packages/ripple-storybook/.storybook/public/img/image-landscape-s.jpg
