+++ packages/ripple-storybook/.storybook/public/img/image-light.jpg
