+++ packages/ripple-storybook/.storybook/public/img/ripple-logo.png
