+++ packages/ripple-storybook/.storybook/utils/svgPlaceholder.js
