+++ packages/ripple-storybook/.storybook/utils/withBackground.js
