+++ packages/ripple-storybook/package.json
@@ -1,0 +1,43 @@
+{
+  "name": "ripple-storybook",
+  "version": "0.0.1",
+  "private": true,
+  "packageManager": "pnpm@7.23.0",
+  "license": "Apache-2.0",
+  "scripts": {
+    "start": "start-storybook -p 6006",
+    "build": "build-storybook",
+    "test": "test-storybook --stories-json",
+    "ci": "concurrently -k -s first -n \"SB,TEST\" -c \"magenta,blue\" \"npx http-server storybook-static --port 6006 --silent\" \"wait-on tcp:6006 && pnpm test\"",
+    "chromatic": "npx chromatic --storybook-base-dir ../ripple-ui-core --storybook-config-dir ../ripple-ui-core"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.20.12",
+    "@originjs/vite-plugin-content": "^1.0.2",
+    "@storybook/addon-a11y": "^6.5.15",
+    "@storybook/addon-actions": "^6.5.15",
+    "@storybook/addon-essentials": "^6.5.15",
+    "@storybook/addon-interactions": "^6.5.15",
+    "@storybook/addon-jest": "^6.5.15",
+    "@storybook/builder-vite": "^0.2.7",
+    "@storybook/jest": "^0.0.10",
+    "@storybook/manager-webpack4": "^6.5.15",
+    "@storybook/test-runner": "^0.9.2",
+    "@storybook/testing-library": "^0.0.13",
+    "@storybook/vue3": "^6.5.15",
+    "@testing-library/vue": "^6.6.1",
+    "@types/jest": "^29.2.5",
+    "@vitejs/plugin-vue": "^3.1.2",
+    "babel-loader": "^8.3.0",
+    "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1",
+    "ts-jest": "^29.0.4",
+    "typescript": "^4.9.4",
+    "vite": "3.2.5",
+    "yaml": "^2.2.1"
+  },
+  "dependencies": {
+    "concurrently": "^7.6.0",
+    "vite-svg-loader": "^4.0.0"
+  }
+}
