+++ packages/ripple-ui-core/src/components/icon/Icon.tokens.stories.mdx
@@ -43,7 +43,6 @@ export const IconSpriteSheetOnDemandTemplate = (args) => ({
 
 <Meta
   title='Base Styles/Icons'
-  parameters={{viewMode: 'docs'}}
   argTypes={{
     colour: {
       control: { type: 'select' },
