+++ packages/ripple-ui-core/src/components/icon/icon.stories.mdx
@@ -19,7 +19,7 @@ export const SingleTemplate = (args) => ({
 })
 
 
-<Meta title='Base styles/Icons' parameters={{viewMode: 'story'}}/>
+<Meta title='Base styles/Icons' />
 
 # Icon playground
