+++ packages/ripple-ui-core/stories/colors.stories.mdx
@@ -28,7 +28,6 @@ export const Template = (args) => ({
 <Meta
   title='Base Styles/Colors'
   parameters={{
-    viewMode: 'docs',
     design: {
       type: 'figma',
       url:
