+++ packages/ripple-ui-core/stories/container.stories.mdx
@@ -18,7 +18,7 @@ export const Template = (args) => ({
 
 # Container
 
-<Meta title='Base Styles/Container' parameters={{ viewMode: 'docs'}}/>
+<Meta title='Base Styles/Container' />
 
 <Canvas>
   <Story name="Container" parameters={{
