+++ packages/ripple-ui-core/stories/grid.stories.mdx
@@ -7,7 +7,6 @@ import { Meta, Canvas, Story } from '@storybook/addon-docs'
       type: 'figma',
       url: 'https://www.figma.com/file/jtsOPNk1k4cWnumzkZgclD/Ripple-2.0-PoC?node-id=1286%3A17345',
     },
-    viewMode: 'docs'
   }}
   decorators={[
     () => ({
