--- packages/ripple-ui-core/stories/introduction.stories.mdx
@@ -1,16 +1,0 @@
-import { Meta } from '@storybook/addon-docs'
-
-<Meta
-  title='Introduction'
-  parameters={{ viewMode: 'docs'}}
-/>
-
-<img src="img/ripple-logo.svg" width="150"  />
-
-## Welcome to Ripple Storybook
-
-
-Storybook - https://storybook.js.org/ is used to test Ripple components in isolation and provide a place find out how a given component works.
-It is used when developing new components and modyfying existing ones.
-
-<a href="/docs/base-styles">Base Styles</a>
