+++ packages/ripple-ui-core/stories/layers.stories.mdx
@@ -102,7 +102,7 @@ export const ElevationTemplate = (args) => ({
 
 # Layer Styles
 
-<Meta title='Base Styles/Layer styles' parameters={{viewMode: 'docs'}}/>
+<Meta title='Base Styles/Layer styles' />
 
 ## Border radius
