+++ packages/ripple-ui-core/stories/spacing.stories.mdx
@@ -31,7 +31,7 @@ export const SpacingTemplate = (args) => ({
 
 # Spacing
 
-<Meta title='Base Styles/Layout' parameters={{ viewMode: 'docs'}}/>
+<Meta title='Base Styles/Layout' />
 
 <Canvas>
   <Story name="Spacing" parameters={{
