+++ packages/ripple-ui-core/stories/typography.stories.mdx
@@ -87,7 +87,7 @@ export const ListsTemplate = (args) => ({
 });
 
 
-<Meta title='Base Styles/Typography' parameters={{ viewMode: 'docs'}}/>
+<Meta title='Base Styles/Typography' />
 
 # Lists
