+++ packages/ripple-ui-core/stories/utilities.focus.stories.mdx
@@ -88,7 +88,7 @@ export const FocusTemplate = (args) => ({
 
 # Focus states
 
-<Meta title='Base Styles/Utilities' parameters={{ viewMode: 'docs'}}/>
+<Meta title='Base Styles/Utilities' />
 
 <Canvas>
   <Story name="Focus" parameters={{
