+++ packages/ripple-ui-core/stories/utilities.spacing.stories.mdx
@@ -84,7 +84,7 @@ export const SizeTemplate = (args) => ({
 
 # Utilities
 
-<Meta title='Base Styles/Utilities/Spacing' parameters={{ viewMode: 'docs'}}/>
+<Meta title='Base Styles/Utilities/Spacing' />
 
 ## Margin
